Add sane defaults, including default for M [iqr_milp]
- [ ] Add defaults to preprocess_iqr_milp
- [ ] Pass defaults from preprocess_iqr_milp to iqr_milp (see below)

In iqr_milp, if M is NULL, then we use 10 * standard deviation of the
residuals from the quantile regression of Y on D and X.
We compute the standard deviation of the errors via `quantreg::rq()`.
In preprocess_iqr_milp, we call iqr_milp many times.
So, rather than calling `quantreg::rq` every time during the
preprocess_iqr_milp phase, let's just use the `quantreg::rq` from
preprocess_iqr_milp to determine M.
That is, when using preprocess_iqr_milp, we should not pass NULL to M in
iqr_milp