Fix constraints, fix vtype, fix output [iqr_milp]
I was using `matrix` when I should have been using `diag` and vice
versa, especially with the complementary slackness constraints.
The dual variables, a, were programmed to be binary instead of
continuous -- I fixed that now.
The complementary slackness condition with a and l had 0's on the rhs,
when it should have been 1's.
The answer to the linear program was not being retrieved properly.