Clean up, add progress messages, let *_minus be empty [test_stat]
To be used in quantreg::rq within FUN, Phi_J_minus and other *_minus
matrices are now empty.
Before, they were a column vector of 0s, but this led to singular design
matrices.
With the previous commit, the quantreg::rq function is called
conditionally based on the number of columns, which allows us to let
*_minus matrices be empty without much consequence.
Furthermore, empty matrices have no influence during cbind operation,
which is what we wanted in the first place (TODO).