Run QR if p_D = 0 a.k.a. full-vector inference [test_stat]
In full-vector inference, we concentrate out all the endogeneous
variables.
Therefore, it is meaningless to solve an IQR MILP problem when there are
no D's and Phi's, i.e., p_D = 0.
Instead, we will just run a quantile regression when p_D = 0.
In the case of full-vector inference, we run a QR of the
concentrated-out Y on X (no D, no Phi).