Get bounds depending on qr$summary [line_confint]
`summary.rq` returns different formats when applied to a quantile
regression, i.e., an object of the `rq` class.
Sometimes, it returns the lower and upper bound.
Other times, it returns the standard error and other relevant info.
So, depending on the output, we will need to either compute the bounds
ourselves or use the bounds as given by `summary.rq`.