Change `M` to be fcn of `max(abs(resid))` [*iqr_milp,miqcp_proj]
Before, the big `M` constant was a factor of the sd of the residuals. It
makes more sense to make it a factor of the largest absolute residual.
Currently, this factor is 5, but in the future, we should make this
smarter.
We should choose this factor to be large (e.g., 10) if the largest and
second-largest residuals are close together.
Else, if the largest residual is far away from the second-largest
residual (i.e., the largest residual is likely an outlier), then the
factor needn't be large (e.g., the factor can be 2).