Run QR if p_D = 0 [test_stat]
When running line_confint() in the setting where there is only one
endogeneous variable, the call to test_stat() in the parallelized
for-loop resulted in problems. Namely, after concentrating out the
endogeneous variable in the for-loop, the test_stat function
couldn't get past the short-iqr part of the code.
Note the preprocess_iqr_milp that is being called when
computing the short-iqr further calls the quantreg::rq since after
concentration, p_D = 0.
This level of depth (line_confint -> test_stat in for loop ->
preprocess_iqr_milp in short-iqr -> rq (if p_D = 0)) might be difficult to handle.

So, instead of calling preprocess_iqr_milp in the test_stat code, I just
called rq() when p_D = 0.
Now, the roadmap is:
line_confint -> test_stat in for loop -> rq (if p_D = 0)