Add messages for debugging [line_confint]
The error I was getting was:
Error in { : task 1 failed - "non-numeric argument to binary operator"

I added messages in the while loop inside the foreach loop and analyzed
the contents of the outfile to see where it was breaking.
It looks like it was breaking when we compared the old and current
p-value to see if the p-value flattens.
The reason this doesn't work is that the old-p-value might take on the
value of "skipped".
This can't be compared with a numeric value for current-p-value in case
the current step isn't skipped.

The solution will be to only rewrite old-p-value with a numeric
variable, i.e., don't store "skipped".
Put differently, we'll only store the p-value of the latest *successful*
step.