Fix: Save the last *successful* step's p-val in old p-val [line_confint]
See previous commit message for the error description.

This fix prevents us from replacing old_p_val with "skipped".
This is a problem if current_p_val is numeric (i.e. the current step
succesfully ran).
The reason this is a problem is that we need to compare old_p_val with
current_p_val to check if the p-value flattens.

So, this fix ensures that p-val only contains the p-value of the most
recent step that successfully ran without hitting the time limit.