Create a new \code{distance_method} [random_walk_subsample]
The first distance method uses the norm sum of the xi's.
Our reasoning for this choice was that the xi's characterize how far we
are away from satisfying the FOCs.

The second distance method that we introduce in this commit computes the
distance between each subsample in the random walk and a reference
subsample. The intent was that this reference subsample is a subsample
inside the FOC polytope of the global optimal solution. The reasoning
for this method is that the global FOC polytope is not so far off form
the rest of the polytopes. However, when we tried to accept/reject using
this distance method, the subsamples that result from this MCMC are not
inside the FOC polytope of the proposed active basis. I.e., our idea
that the FOC polytopes of the proposed active basis is close to the FOC
polytope of the globally optimal active basis is incorrect. In fact,
these subsamples aren't even inside the FOC polytope of the actual
active basis. This, however, may be due to the fact that for our
particular testing data, the subsample returned by
`find_subsample_in_polytope` function was unfortunately not in the FOC
polytope.
So ultimately, this method may not be so useful compared to the existing
method of `distance_method = 1`.