refactor(rwalk_subsample)!: split aux and subsamples
Before, I ran two for-loops, one for the subsamples and another for the
aux variables.
The only difference between the two for-loops was that in the aux
for-loop, I was computing Q(x | beta_star) in addition to Q(x | beta_hat).

So, I decided only to use one for-loop.
To propose the subsamples, I would use:
```r
rwalk_subsample(
  h = h_star,
  initial_subsample = rounded_center,
  iterations = K,
  ...,
)
```
If I wanted to work with aux variables, I would use:
```r
rwalk_subsample(
  h = h_opt,
  initial_subsample = rounded_center_opt,
  iterations = Kprime,
  h_alt = h_star, # this means we will compute Q(x | beta_star)
  ...
)
```