fix(mcmc-subsample): avoid numerical problems using a while loop
In `rwalk_subsample`, it is possible for exp{-\gamma * FOCviolation} to
be really small (i.e., FOC violation is super big.).
For such cases, R may replace the true value of
exp{-\gamma * FOCviolation} with 0.
This raises computational concerns (e.g., division by 0).
How do we proceed?

Well, if we run into such a case, it means that the proposed subsample
is outside the FOC polytope, so it contributes 0 to the sum that we are
trying to unbiasedly target.
Thus, we can safely ignore this iteration and try again.

This solution is implemented in this commit.
I wrap the contents of the for-loop in a while-loop that breaks if we
don't have any numerical problems.
If we do have numerical problems, we try the iteration again.