feat(center): omit `h` from `foc_center` and use coefficients
This is useful when we don't want to use the active basis but we want to
use the coefficients.
For example, rather than solving a big MILP to get the active basis, we
can solve the MILP on a subset of the data.
Suppose the original data is of size n and the subset is of size n'.
But this will yield an active basis is a binary vector of size n' with
p number of 1's.
How would we then convert this into an active basis that is a binary
vector of size n?
To avoid having to answer this problem, we can simply use the
coefficients associated with the active basis of size n' in
`foc_center`.
However, the program used in `foc_center` will continue to use n as the
size of the problem.