fix(alt_sample): correctly sample; use integers
`sample(x)` when `x` is of length 1 is akin to `sample(1:x)`, which is
not what we want.
In our setting, when `x` is of length 1, we just want to return `x` --
we don't want to treat `x` as some sort of max count. We want to ensure
`x` is an index.