fix(mcmc-subsample): use correct proposal scheme and proposal ratio
In this commit, I propose a direction uniformly at random from the
available directions.
Then, I choose a one-step neighbor unifromly at random in that
direction.
Such a sampling scheme is consistent with the proposal distribution we
have in mind.
Before, the sampling scheme did not follow the desired proposal
distribution.

Before, I wasn't properly accounting for the proposal ratio.
I mistakenly lumped it with the target probability, but I don't think
this is correct.
Here, I explicitly consider log_Q_star and log_Q_current which is the
proposal of Dstar given D_{k - 1} and proposal of D_{k - 1} given Dstar,
respectively.